Editor: add window with terrain list

+++ tools/Map editor/Map editor/Map.h
@@ -29,7 +29,7 @@ class Map {
   void draw(std::pair<int, int> offset);
 
  private:
-  size_t _dimX{0}, _dimY{0};
+  int _dimX{0}, _dimY{0};
   int _textureWidth{0}, _textureHeight{0};
 
   static const int TILE_SIZE = 2;
